Create 1 .Java Graal
This code defines a simple Java class with a main method that calls the add method and prints the result. The add method is annotated with the @CompileStatic annotation, which tells the GraalVM compiler to statically compile the method. The @Fold annotation tells the compiler to perform constant folding, which can further improve the performance of the method.

To compile this code with GraalVM, you can use the native-image command as shown above. This will create a native executable that contains the compiled code for the add method.

Java Graal is a technology that allows Java code to be compiled ahead-of-time (AOT) into native machine code, resulting in faster startup and execution times. Graal is the name of the JIT (just-in-time) compiler used in the Java Virtual Machine (JVM), and the Java Graal technology uses this compiler to generate native code for Java applications.

Java Graal is implemented as a native-image utility that is included with the GraalVM distribution. It can be used to compile a Java application into a standalone native executable that can be run on a machine without a JVM installed.

Java Graal is particularly useful for applications that need to start and run quickly, such as serverless functions or microservices. By compiling the code ahead-of-time, the application can avoid the overhead of JVM startup and interpretation of bytecode, leading to faster execution times