Create Test containers
This code creates a PostgreSQLContainer object, which is annotated with the @Rule annotation. This tells JUnit to create and start the container before running the test method, and to stop and remove the container after the test method has completed. The getJdbcUrl, getUsername, and getPassword methods can be used to obtain the connection details for the PostgreSQL instance.he add method.
Testcontainers is a Java library that allows developers to run test environments in Docker containers as part of their integration tests. It is designed to make it easy to set up and tear down containerized test environments, and to provide convenient access to the services running in those environments.

Testcontainers provides support for a wide range of services, including databases, message brokers, and web servers. It also provides a number of utility classes for common tasks such as waiting for a service to become available or generating random data.