Create 3.Spring boot application
This code defines a simple Spring Boot application with a single @RestController that has a single @GetMapping for the /hello endpoint. When the application is run and you send a GET request to /hello, the controller will return the string "Hello, World!".

To run this application, you will need to have the Spring Boot framework on your classpath. You can then use the SpringApplication.run method to start the application