Create connecting springboot app to databse
This code defines a Spring Boot application with a PostgreSQLContainer bean that is started as part of the application's context. The container is then injected into the application using the @Autowired annotation.

The code also defines a CommandLineRunner bean that is run when the application starts. This bean uses the PostgreSQLContainer to obtain a JDBC connection to the database and can use the connection to execute queries or perform other database operations.